week3 assignment
Create a class called "Car" that includes the following attributes: a make (string), model (string), and speed/mph (int).

Your program's "Main" should ask the user to enter the speed limit. The user should then be prompted to enter information for 3 different cars (make, model, and speed).

If a car is under the speed limit, the Console should display "OK" for that car, along with that car's information. If a car is above the speed limit, the program should calculate the number of demerit points. For every 5mph above the speed limit, 1 demerit point should be incurred. Display the car's information, how many demerit points it received, and if more than 10, display "License Suspended"